test internet access
