+++ lambdas/package.json
@@ -12,7 +12,8 @@
   "dependencies": {
     "@middy/core": "^3.4.0",
     "@middy/http-json-body-parser": "^3.4.0",
-    "aws-sdk": "^2.1234.0"
+    "aws-sdk": "^2.1234.0",
+    "axios": "^1.1.3"
   },
   "devDependencies": {
     "@serverless/typescript": "^3.0.0",
