+++ lambdas/src/functions/buildPackages/mock.json
