+++ lambdas/src/functions/index.ts
@@ -1,1 +1,1 @@
-export { default as hello } from './hello';
+export { default as hello } from './buildPackages';
