+++ lambdas/src/libs/types.ts
@@ -1,0 +1,11 @@
+export interface Package {
+    name: string;
+    full_name: string;
+    slug: string;
+    desc: string;
+    homepage: string;
+    version: string;
+    installs: number;
+    maintainer: string;
+    thumb_image_url: string;
+}
